Adds Plotting API (#354)
* Add new PlotterMixin

* Moved plotting function in examples/plotting.py to new subpackage gtda/plotting

* Add PoC plot func in VietorisRips
