+++ doc/index.rst
@@ -341,6 +341,43 @@ Binary image filtrations
    pipeline.make_pipeline
 
 
+:mod:`gtda.utils`: Validation
+=============================
+
+.. automodule:: gtda.utils
+   :no-members:
+   :no-inherited-members:
+
+.. currentmodule:: gtda
+
+.. autosummary::
+   :toctree: generated/
+   :template: function.rst
+
+   utils.check_diagram
+   utils.validate_params
+
+
+:mod:`gtda.plots`: Plotting
+=============================
+
+.. automodule:: gtda.plots
+   :no-members:
+   :no-inherited-members:
+
+.. currentmodule:: gtda
+
+.. autosummary::
+   :toctree: generated/
+   :template: function.rst
+
+   plots.plot_point_cloud
+   plots.plot_diagram
+   plots.plot_betti_curves
+   plots.plot_betti_surfaces
+   plots.plot_landscapes
+
+
 :mod:`gtda.meta_transformers`: Convenience pipelines
 ====================================================
 
@@ -357,20 +394,3 @@ Binary image filtrations
    meta_transformers.EntropyGenerator
    meta_transformers.BettiCurveGenerator
    meta_transformers.LandscapeGenerator
-
-
-:mod:`gtda.utils`: Validation
-=============================
-
-.. automodule:: gtda.utils
-   :no-members:
-   :no-inherited-members:
-
-.. currentmodule:: gtda
-
-.. autosummary::
-   :toctree: generated/
-   :template: function.rst
-
-   utils.check_diagram
-   utils.validate_params
