+++ gtda/__init__.py
@@ -1,4 +1,5 @@
 from ._version import __version__
 
 __all__ = ['homology', 'time_series', 'graphs', 'diagrams', 'images',
-           'point_clouds', 'externals', 'meta_transformers', '__version__']
+           'point_clouds', 'externals', 'meta_transformers', 'plotting',
+           '__version__']
