+++ gtda/base.py
@@ -1,5 +1,5 @@
 """Implements a TransformerResamplerMixin for transformers that have a resample
-method."""
+method and TransformerPlotterMixin for transformers that have a plot method."""
 # License: GNU AGPLv3
 
 
@@ -18,6 +18,7 @@ class TransformerResamplerMixin:
         ----------
         X : ndarray of shape (n_samples, ...)
             Input data.
+
         y : None
             There is no need for a target in a transformer, yet the pipeline
             API requires this parameter.
@@ -26,6 +27,7 @@ class TransformerResamplerMixin:
         -------
         Xt : numpy array of shape (n_samples, ...)
             Transformed input.
+
         """
         # non-optimized default implementation; override when a better
         # method is possible for a given clustering algorithm
@@ -46,6 +48,7 @@ class TransformerResamplerMixin:
         ----------
         X : ndarray of shape (n_samples, ...)
             Input data.
+
         y : ndarray of shape (n_samples,)
             Target data.
 
@@ -53,6 +56,10 @@ class TransformerResamplerMixin:
         -------
         Xt : ndarray of shape (n_samples, ...)
             Transformed input.
+
+        yr : ndarray of shape (n_samples, ...)
+            Resampled target.
+
         """
         return self.transform(X), self.resample(y, X)
 
@@ -65,6 +72,7 @@ class TransformerResamplerMixin:
         ----------
         X : ndarray of shape (n_samples, ...)
             Input data.
+
         y : ndarray of shape (n_samples,)
             Target data.
 
@@ -72,7 +80,70 @@ class TransformerResamplerMixin:
         -------
         Xt : ndarray of shape (n_samples, ...)
             Transformed input.
+
         yr : ndarray of shape (n_samples, ...)
             Resampled target.
+
         """
         return self.fit(X, y, **fit_params).transform_resample(X, y)
+
+
+class PlotterMixin:
+    """Mixin class for all plotters in giotto-tda."""
+
+    def fit_transform_plot(self, X, y=None, sample=0, **plot_params):
+        """Fit to data, then transform and plot a sample in the input
+        collection. Return the transformed sample input.
+
+        Parameters
+        ----------
+        X : ndarray of shape (n_samples, ...)
+            Input data.
+
+        y : None
+            There is no need for a target in a transformer, yet the pipeline
+            API requires this parameter.
+
+        sample : int
+            Sample to be plotted.
+
+        **plot_params
+            Optional plotting parameters.
+
+        Returns
+        -------
+        Xt : numpy array of shape (1, ...)
+            Transformed input sample.
+
+        """
+        self.fit(X, y)
+        Xt = self.transform_plot(X, sample=sample, **plot_params)
+        return Xt
+
+    def transform_plot(self, X, sample=0, **plot_params):
+        """Transform and plot a sample in the input collection.
+        Return the transformed sample input.
+
+        Parameters
+        ----------
+        X : ndarray of shape (n_samples, ...)
+            Input data.
+
+        y : ndarray of shape (n_samples,)
+            Target data.
+
+        sample : int
+            Sample to be plotted.
+
+        plot_params : dict
+            Optional plotting parameters.
+
+        Returns
+        -------
+        Xt : ndarray of shape (n_samples, ...)
+            Transformed input sample.
+
+        """
+        Xt = self.transform(X[[sample]])
+        self.plot(Xt, sample=sample, **plot_params)
+        return Xt
